% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatisticsshiny-package.R
\docType{data}
\name{commodities_short_display}
\alias{commodities_short_display}
\title{commodities_short_display}
\description{
Internal dataset for commodity codes (1,363 codes).
}
\keyword{datasets}
