% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsshinycountryprofiles-package.R
\docType{package}
\name{otsshinycountryprofiles-package}
\alias{otsshinycountryprofiles}
\alias{otsshinycountryprofiles-package}
\title{otsshinycountryprofiles: Open Trade Statistics Country App}
\description{
Production grade Shiny app to display the country profiles for tradestatistics.io.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tradestatistics/visualization-with-shiny}
  \item Report bugs at \url{https://github.com/tradestatistics/visualization-with-shiny/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{m.vargas.sepulveda@gmail.com}

}
\keyword{internal}
