% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatisticsshiny-package.R
\docType{package}
\name{tradestatisticsshiny-package}
\alias{tradestatisticsshiny}
\alias{tradestatisticsshiny-package}
\title{tradestatisticsshiny: Open Trade Statistics Dashboard}
\description{
This is meant as a production grade Shiny app and not as end user app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tradestatistics/visualization-with-shiny}
  \item Report bugs at \url{https://github.com/tradestatistics/visualization-with-shiny/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

}
\keyword{internal}
